Run bundle commands in the app user's login shell
It's a (maybe stupid?) goal to try and run all commands as the application user, even if logging on as another user.  Unfortunately, to make proper use of SSH-agent forwarding, git still needs to be run as the original user, but bundle and other ruby commands should not.

Doing this has the big advantage that deciding which tools are available to the app (such as which ruby) can be done in the application user's profile.  Other environment variables (such as the rails environment, or maybe remote database credentials and similar) will also be available, and can be kept out of deployment scripts.