Change the way environment variables are set to allow spaces in values and references to other environment variables.
This should now allow setting variables like: `PATH=/new/path:$PATH` or `NAME=Tom Ward`.

N.B. To get this to work, the way in which the environment is set has changed.  Previously the application's `.profile` had an if statement like `if [ -s "$HOME/.env" ] ... fi` which loaded the environment.  It now uses a dedicated `.recap` script, which is loaded in `.profile` using `. $HOME/.recap".  To upgrade existing apps, remove the previous if statement, and run cap bootstrap again.  Hopefully if changes need to be made in the future they can be done by simply updating the `.recap` file, leaving `.profile` untouched.