The index of the documentation should be the `Recap` module.
Having `index.rb` was a bit of a hack, and with a bit of massaging we
can convince Rocca to document the files in `lib` as if they were the
root of the application.

In order to make the directory index URL work, we also generate a simple
`index.html` file which redirects to `recap.html`, which is the start of
the documentation.