Mocha::Expectation#in_sequence with no Mocha::Sequence has no effect.
A number of tests setup expectations and attempt to constrain them to
occur one after the other using Mocha::Expectation#in_sequence, but
without supplying any instances of Mocha::Sequence. This means that the
expectations are not constrained to occur in the specified order.

I've added a sequence so that the intended constraint is enforced.

I've also added a Mocha issue [1] to highlight this potential for confusion.

[1] https://github.com/floehopper/mocha/issues/79