Use expectations to check for calls to Kernel#abort.
Previously we were checking that a SystemExit exception was being
raised, but this meant that the error message was send to stderr.

Note that you have to be careful to set the expectation on the correct
capistrano namespace object, which makes this a little awkward, but I
think it's ok.