Add some more helpful output if the user or group isn't found.
While running the `bootstrap` task is the easiest way of getting your server into the right state for deploying using recap, it's not the only way, and there could be various reasons why you either don't want to run that task, or it cannot complete automatically.

The preflight tasks are designed to check at every deployment that the server is in the correct state, and advise you of how you can bring it up to that state if it is not.

In the worst case (not running `bootstrap` at all), the preflight checks should prompt you one at a time with how to fix the server setup, until the project is deployable. A bit liking fixing one test at a time.

Fixes #46.