If `deploy:setup` fails, we should be able to run it again.
If the deploy setup failed for some reason (e.g. bundle install failed because the right version of Ruby wasn't in the PATH), the git repository would be left in the filesystem. Subsequent runs of `deploy:setup` would then fail for a different reason, because it was trying to re-clone the git repository into the deploy location.

This change ensures that if `deploy:setup` fails for any reason, the repository is removed, and running `deploy:setup` again will result in the same error until the specific application problem is fixed.

Fixes #48.