The application user needs to be able to execute `bundle`.
The preflight check was checking that the bundle command was available
for the deploying user; whereas it is actually the application user
that needs access. This meant that some of the cucumber features were
incorrectly suggesting that bundler was not installed and were
consequently failing. In fact bundler was installed, but just not
for the deploying user. We've tried to make the error message even more
specific to make it as useful as possible.