Host documentation at gofreerange.com/recap/docs
It's quite simple to upload the files to our server; in the virtual host
for Apache, I've added the following configuration:

    AliasMatch ^/recap/docs(.*)$ /home/freerange/docs/recap$1
    <Directory "/home/freerange/docs/recap">
      PassengerEnabled off
      AllowOverride all
      allow from all
      Options +Indexes
    </Directory>

Since with this change we wouldn't use github to host the docs, we don't
need the documentation in git, so I've also removed the files and
ignored the `doc` directory. If this is acceptable, we can also remove
the gh-pages branch.