Ensure ~/.recap can be sourced from anywhere
Sourcing ~/.recap from outside the home directory fails to generate the
.recap-env-export file correctly because it expects to find .env in the
current directory.

This fix ensures we read ~.env and generate ~/.recap-env-export
correctly.

I came across this problem when using the whenever gem to generate my
crontab.  It generate a command similar to /bin/bash -l -c 'my-command'
which I tried to run from the command line.  Running it in a directory
outside of ~ fails due to the problem mentioned above.  *NOTE* The key
part of that command is '-l' which tells bash to load ~/.profile (among
other things).